+++ src/main/lrphotos.lrdevplugin/SearchExtraPhotosInPhotos.lua
@@ -73,19 +73,16 @@ local function sendPhotosToApp(action, maintenanceDir, catalogNames, albumName)
 
     local command = ""
     LrFileUtils.delete(maintenanceDir .. "/fromPhotos/photos.txt")
-    logger.trace("1")
     local exePathDev = _PLUGIN.path .. "/" .. "PhotosMaintenance/PhotosMaintenance.app"
     local exePathProd = _PLUGIN.path .. "/" .. "PhotosMaintenance.app"
     logger.trace("exePathDev=" .. tostring(exePathDev))
     if LrFileUtils.exists(exePathDev) then
-        logger.trace("2")
         command = "osascript \"" .. exePathDev .. "\""
                 .. " " .. action
                 .. " \"" .. maintenanceDir .. "\""
                 .. " \"" .. catalogNames .. "\""
                 .. " \"" .. albumName .. "\""
     else
-        logger.trace("3")
         command = "osascript \"" .. exePathProd .. "\""
                 .. " " .. action
                 .. " \"" .. maintenanceDir .. "\""